+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 ## 4.3 - 12/05/2018
 ### Added
 - Regex for security on all form
+- More back-end security : send message after checking for XSS
 ## 4.2 - 11/05/2018
 ### Added
 - Class Discord, for a cleaner code