+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+## 4.4 - 13/05/2018
+### Modified
+- CSS on the channel list page : button for logout and go back home
 ## 4.3 - 12/05/2018
 ### Added
 - Regex for security on all form
-- More back-end security : send message after checking for XSS
+- More back-end security : send message after checking for XSS, checking user input length
 ## 4.2 - 11/05/2018
 ### Added
 - Class Discord, for a cleaner code